# CUDA : faster k-quant dot kernels