--- a/ggml-cuda.cu
+++ b/ggml-cuda.cu
@@ -168,7 +168,7 @@ static_assert(sizeof(block_q6_K) == sizeof(ggml_fp16_t) + 13*QK_K/16, "wrong q6_
 #endif
 
 #ifndef K_QUANTS_PER_ITERATION
-#define K_QUANTS_PER_ITERATION 1
+#define K_QUANTS_PER_ITERATION 2
 #endif
 
 static __global__ void add_f32(const float * x, const float * y, float * dst, const int k) {