After the CUDA inference speed up in PR #1827 I got motivated to improve the performance of the k-quants, and this PR is the result.

The following table compares token prediction times on current master (i.e., after merging #1827) and this PR for all k-quantization levels. Shown is the average of 3 runs in milliseconds predicting 128 tokens on an RTX-4080 GPU using the following command:
```
./bin/main -m ../models/7B/qXk.bin -p "I believe the meaning of life is" -c 2048 -n 512 --ignore-eos -n 128 -s 1234 -t 1 -ngl 100
```

|Quantization|Master|This PR| Speedup|
|--:|--:|--:|--:|
|Q2_K | 11.94 | 10.62 | 12.4% |
|Q3_K_S| 15.35 | 11.68 | 31.4% |
|Q4_K_S| 12.09 | 10.98 | 10.1% |
|Q5_K_S | 13.53 | 12.03 | 12.5% |
|Q6_K | 14.51 | 13.33 | 8.9% |
|Q4_0 | 12.40 | - | - |

I have added `Q4_0`, not touched by this PR, for comparison. It is now more than 10% slower than `Q4_K`, bit it should be faster as it needs to do less work, so there is a potential for improvement there (on Metal, `Q4_0` is ~10% faster than `Q4_K`).

I have moved away from using a templated implementation. There are subtle differences between the different k-quants in what is the best way to spread the threads in a thread group within a quantization block, and these variations are difficult to capture in a template. The result is ~100 extra lines of code compared to the templated implementation, which I think is worth it, given the increase in performance.