@JohannesGaessler 2023-06-14
I did some performance testing: 

| GPU      | Model  | ms/t master | ms/t PR |
|----------|--------|-------------|---------|
| RTX 3090 | q2_K   |       17.24 |   16.49 |
| RTX 3090 | q3_K_S |       21.31 |   18.42 |
| RTX 3090 | q4_K_S |       17.74 |   17.11 |
| RTX 3090 | q5_K_S |       18.84 |   17.98 |
| RTX 3090 | q6_K   |       18.76 |   17.62 |
| RTX 3090 | q4_0   |       15.76 |       - |
| GTX 1070 | q2_K   |       91.87 |  127.91 |
| GTX 1070 | q3_K_S |       91.59 |   81.96 |
| GTX 1070 | q4_K_S |      129.12 |  134.02 |
| GTX 1070 | q5_K_S |      144.76 |  136.66 |
| GTX 1070 | q6_K   |       81.73 |  143.29 |
| GTX 1070 | q4_0   |       62.61 |       - |

On my RTX 3090 the new kernels are consistently faster. q4_0 is still faster though (maybe you didn't compile with `LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_DMMV_Y=2`?). On my GTX 1070 the performance is only better for two out of the five quantization formats. For q2_K and q6_K the performance is significantly worse.