@ikawrakow 2023-06-15
@JohannesGaessler Thank you for taking the time to look into this PR and posting results for different cards. Do you have a hypothesis why the new version tuns slower on the 1070?

You are right, i did not use `LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_DMMV_Y= 2` for the `Q4_0` result. Using this, `Q4_0` token prediction time becomes 11.5 ms on my 4080 card, so still slower than `Q4_K_S`.