@JohannesGaessler 2023-06-15
>It would be great if the performance of q2_K and q6_K on older GPUs can be fixed, but if you don't have a Pascal GPU for testing, I think it is ok prioritizing the performance on newer architectures.

I personally would not merge an optimization that causes a small performance uplift for some users but a large performance malus for other users but I am neither the project lead nor the developer of k-quants so I'll leave the decision up to @ikawrakow 

>Do you have a hypothesis why the new version tuns slower on the 1070?

I don't. GPU performance optimizations depend very strongly on the hardware details but the difference between an RTX 3090 and a GTX 1070 is quite large. It oftentimes comes down to something like the number of warps that fit into the available registers on a streaming multiprocessor and it's much easier to just tweak parameters and test performance than to try to understand what's happening on the hardware side. It's also possible that you are now using some feature that just has bad performance on older cards.