@ggerganov 2023-06-15
I'm also leaning towards improving the performance on newer devices and optionally having compile flags / macros for better performance on older hardware **given that this does not complicate or duplicate the code too much** (i.e. similar to the `GGML_CUDA_DMMV_X`, `GGML_CUDA_DMMV_Y` defines).

But before making any decision, let's wait for one or two more reports that confirm the degraded performance on older cards, just in case