@ikawrakow 2023-06-15
I did get my hands on a box with a GTX-1660 (released Oct 2019). `Q3_K, Q4_K` and `Q5_K` are faster with this PR, `Q2_K` and `Q6_K` are slower. So, I guess, I'll do something with `ifdef`s for `Q2_K` and `Q6_K`.

If I remember correctly, back then the 1660 was basically the cheapest option if one wanted to have a discrete GPU but didn't want to pay the premium Nvidia added with the introduction of the 2000 series. The card has 6 GB of VRAM and the 4-bit models fail to run on it with `-ngl 100` on current master. They run with `-ngl 32`, but performance is lower compared to what we had before #1827 was merged, so it seems we merged a PR without being too much worried about the experience of users on older hardware.    

I also got my hands on an even older PC gathering dust in the corner. It has a 1060 card in it, so I wanted to experiment with that. But after fighting with Nvidia drivers and CUDA toolkits for a while, I eventually gave up. Are we aware that none of the CUDA stuff will work with the default packages on Ubuntu 20.04 LTS? I any case, while at it, I ran `Q4_K_S` on the CPU (i9-9960X, released 2018) and it was faster than the numbers @JohannesGaessler posted for the 1070 card.