@johnson442 2023-06-16
Some more data from the type of hardware in question - 1070ti and i5-8600K:

|	|Master |PR | CPU (-t 5)|
|-------|-------------|---------|-----------|
| 7B 	|		|	|	|
| Q2_K 		|84.92		|116.34	|	83.68	|
| Q3_K_S 	|83.07		|75.21	|	86.10	|
| Q4_K_S 	|120.97		|122.55	|	102.77	|
| Q5_K_S 	|136.51		|127.96	|	122.62	|
| Q6_K	|74.18		|131.62	|	143.83	|
| Q4_0	|57.50		|-	|	103.09  |
|13B (-ngl 24 -t 5)	|		|	|		|	
| Q6_K	|196.96		|255.84	|	273.61  |

I've included CPU only (with 5 threads on 6 cores) to compare, as well as 24 layers of a 13B model on the GPU (max in 8GB of VRAM).

Forgive me if its explained elsewhere, but is there a reason Q6_K is faster than the other new quants on this GPU?