@maddes8cht 2023-06-16
With my rtx 3060 using the #1827 build performance got actually worse when using the "extra layers" with 13b models.
Generation times got about 40%  slower, and prompt eval time raised about 25 times up ( from something like 80 ms/token to 2000 ms/token)
Going back to -ngl 40 with 13b models braught me back a lot of performance.

How about making block sizes configurable at run time?
And maybe have a small helper app that tests out optimal parameters?