@ikawrakow 2023-06-16
> Forgive me if its explained elsewhere, but is there a reason Q6_K is faster than the other new quants on this GPU?

The `Q6_K` quantization is simpler. The weight `w` is obtained from the quant `q` as `w = d * s * q`, where `d` is the scale of the super-block and `s` is the scale of the block inside the super-block in which `q` is located. In `Q6_K` the scale is 8 bit, so it can be taken and used directly. `Q3_K` is similar, but there the scale is 6 bit, so one needs more compute (bit manipulations) to  recover it from the densely stored data. For the other k-quants we have `w = d * s_q * q - m * s_m`, where now `m` is another scale for the super-block and `s_q, s_m` are per-block 6-bit parameters (4-bit in the case of `Q2_K`). As a result one needs way more bit fiddling to recover the scales and the quants.

When the computation becomes very fast as it is the case on the latest generation Nvidia cards, memory bandwidth becomes important, and this results in `Q6_K` being slower than the others on my RTX-4080. To give more specific numbers, we have `Q6_K` taking `~13.5` ms. If I put `return` as the very first statement in  the dot product kernel, the computation completes in `~3.5` ms/token. I.e., the dot product takes `~10` ms. The `Q6_K` 7B model is 5.5 GB, so this means `~550` GB/s are being read from memory while the dot products are being computed. The RTX-4080 spec claims 768 GB/s memory bandwidth, but that would be under ideal conditions with all threads/warps operating in sync and fetching data from memory in the optimum way. We don't have that here as we are reading single bytes and jumping around within the data of a super-block to fetch scales/quants, so my guess is that 550 GB/s is as good as it gets (or at least very close to). In comparison, `Q4_K_S` runs in `~11` ms, so `~7.5` ms for dot products. The model is 3.8 GB, so ~500 GB/s are being read from VRAM while running inference with a `Q4_K` model.