@JohannesGaessler 2023-06-16
@maddes8cht I saw some people complain about how with the newest NVIDIA drivers the program no longer crashes when you run out of VRAM. Instead data from VRAM gets moved to RAM and you get terrible performance. And it's a "feature" that cannot be disabled.