@ikawrakow 2023-06-16
I put some effort into improving the performance of `Q2_K` and `Q6_K` on older GPUs, while preserving performance on newer cards. The oldest GPU I have available is a GTX-1660 released in 2019, so I would appreciate if someone tested on the the GTX-1000 series. To deal with older cards, I have added `LLAMA_CUDA_KQUANTS_ITER` to `CMakeLists.txt` and `Makefile`. It is set by default to `2` (best for RTX-4080). On older cards performance is better with `LLAMA_CUDA_KQUANTS_ITER=1`.

Here are the per token prediction times for 128 tokens using the 7B model after the last commits on a GTX-1660 using `LLAMA_CUDA_KQUANTS_ITER=1` (performance on the RTX-4080 is the same as given in the initial description of the PR):

|Quantization|Master|This PR| Speedup|
|--:|--:|--:|--:|
|Q2_K | 52.2 | 47.4 | 10.1% |
|Q3_K_S| 65.6 | 60.3 | 8.8% |
|Q4_K_S| 68.0 | 65.5| 3.8% |
|Q5_K_S | 74.4 | 68.5 | 8.6% |
|Q6_K | 63.0 | 62.5 | 0.9% |
|Q4_0 | 50.6 | - | - |
|Q5_1| 87.6 | - | - |

The GTX-1660 has 6 GB of VRAM, so only `Q2_K` and `Q3_K` fit completely on the GPU (including KV cache) and are run with `-ngl 100 -t 1`. `Q4_K` and `Q5_K` were run with `-ngl 32 -t 4`, `Q6_K` with `-ngl 29 -t 4`, and `Q5_1` with `-ngl 30 -t 4`.